Reorganize playbook dependencies
The dependencies can not go to the roles, as we have complex class
based decisions to account for. As in this early step the tasks would
be called multiple times, we just redefine it as in was first: use
install.yml for installation and use install-*.yml just as a short
hand to do the isolated tasks included.